Don't back up original dotfiles if they are equal to repo version

+++ setup.sh
@@ -9,9 +9,9 @@ do
   if [ "$HOME/$file" -ef "$DIR/$file" ]; then
     echo "$HOME/$file already symlinked"
   else
-    if [ -f "$HOME/$file" ]; then
+    if [ -f "$HOME/$file" ] && ! cmp -s "$DIR/$file" "$HOME/$file"; then
       mv -v "$HOME/$file" "$HOME/$file.bak"
     fi
-    ln -sv "$DIR/$file" "$HOME/$file"
+    ln -sfv "$DIR/$file" "$HOME/$file"
   fi
 done
